requirements.txt
Django==4.2.7
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
aiogram==3.2.0
telethon==1.34.0
supabase==2.0.3
psycopg2-binary==2.9.9
cryptography==41.0.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==21.2.0
daphne==4.0.0
asyncio==3.4.3
aioredis==2.0.1
requests==2.31.0